refactor: 작성 페이지 첫 진입 화면 UI 개선
